Merges test/bidirectional_astar.cc with test/astar.cc
